refactor(front): splitted sushi page into two tables
