+++ front/i18n/locales/fr.yaml
@@ -506,7 +506,6 @@ institutions:
     resumeEntryQuestion: Reprendre votre saisie ?
     resumeEntryDesc: Vous avez déclaré la fin de saisie de vos identifiants le {date}. Souhaitez-vous la reprendre ?
     harvest: Moissonnage
-    archiveSwitch: Archiver/Désarchiver
     readyPopup:
       inProgress: Vos identifiants sont encore en cours de saisie. Lorsque ces derniers seront prêts à être moissonnés, déclarez la fin de votre saisie en cliquant sur le bouton ci-dessous.
       completed: Vous avez déclaré la fin de saisie de vos identifiants le {date}. Ces derniers pourront faire partie des prochaines vagues de moissonnage. Si des ajustements sont encore nécessaires, déclarez-le en cliquant sur le bouton ci-dessous.
